Refactor: Move AppInstallPageFactory to a separate file
